config: Add layer to "Ised-CreateJiraIssue" lambda
From version 1.13.0, the "requests" module is no longer part of the AWS
SDK for Python. The Lambda service continues to bundle the requests
module in the AWS SDK until March 30, 2020. [1]

Adding this "layer" to the "Ised-CreateJiraIssue" lambda should
essentially pin the version of the AWS SDK to version 1.12.221, which
contains the "requests" library.

Eventually, we should probably store lambda code in S3 and bundle our
external dependencies, such as "requests" [2]

[1] https://aws.amazon.com/blogs/compute/upcoming-changes-to-the-python-sdk-in-aws-lambda/
[2] https://docs.aws.amazon.com/lambda/latest/dg/lambda-python-how-to-create-deployment-package.html