routing/AELB: Lambda to get next listener priority
When adding a ListenerRule to a Load Balancer, the "Priority" property
is required, and must be unique. Cloudformation has no way of knowing
which number is available, so we use a Lambda (and required IAM
role/permissions) that describes the rules on the target Listener and
returns the next available slot.

This change makes it so that such a Lambda function is created per Load
Balancer to keep its IAM role/permissions from crossing environments.

With that said, the current IAM policy doesn't restrict permissions in a
per-resource manner at the moment, because "DescribeRules" doesn't
support it[1]. This has the effect that the lambdas can describe rules
on any Load Balancer, but in the eventuality that "DescribeRules" gets
support for per-resource permissions, our infrastructure will be ready
and we'll just need to update the policy to restrict permissions to
target Load Balancer.

[1] https://docs.aws.amazon.com/elasticloadbalancing/latest/userguide/load-balancer-authentication-access-control.html