config: Filter AWS resources types down to ones we tag
Originally, the Lambda would iterate over all AWS resources to find the
one that triggered the Config rule. As accounts create more and more
resources, this process takes longer and longer and the lambda ends up
timing out.

This change makes it so that, when asking for the list of resources, we
also pass a list of Resources Types we're interested in (ex:
ec2:volume), which reduces the number of resources returned
significantly.